pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "openai-agents[voice]>=0.9.3",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.5",
]
